move human readable display utilities to `datafusion-common` crate
