+++ datafusion/physical-plan/benches/spill_io.rs
@@ -25,8 +25,8 @@ use criterion::{
     criterion_group, criterion_main, BatchSize, BenchmarkGroup, BenchmarkId, Criterion,
 };
 use datafusion_common::config::SpillCompression;
+use datafusion_common::human_readable_size;
 use datafusion_common::instant::Instant;
-use datafusion_execution::memory_pool::human_readable_size;
 use datafusion_execution::runtime_env::RuntimeEnv;
 use datafusion_physical_plan::common::collect;
 use datafusion_physical_plan::metrics::{ExecutionPlanMetricsSet, SpillMetrics};
